engine: TestStartStopWithCredentials for Windows
